requirements.txt
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
python-decouple==3.8
django-ckeditor-5==0.2.15
